+++ clayrs/content_analyzer/embeddings/embedding_loader/gensim.py
@@ -1,9 +1,8 @@
-from clayrs.content_analyzer.embeddings.embedding_loader.embedding_loader import WordEmbeddingLoader
-
 from gensim import downloader
 import numpy as np
 
 from clayrs.utils.const import logger
+from clayrs.content_analyzer.embeddings.embedding_loader.embedding_loader import WordEmbeddingLoader
 
 
 class Gensim(WordEmbeddingLoader):
