+++ clayrs/utils/__init__.py
@@ -1,3 +1,2 @@
 from .load_content import load_content_instance
 from .report import Report
-from .experiment import ContentBasedExperiment, GraphBasedExperiment
