+++ clayrs/utils/const.py
@@ -1,7 +1,6 @@
 import os
 from clayrs.utils.custom_logger import get_custom_logger
 
-
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
 root_path = os.path.join(THIS_DIR, '../../')
 contents_path = os.path.join(root_path, 'contents/')
